Refactor: split RaftInner into separate file
Moved the `RaftInner` struct and its methods into a separate
`raft_inner.rs` file. `Raft` is a control handle for application,
internally used methods should be provided by `RaftInner`, not
`Raft`.
